+++ elmyra.ip.access.epo/elmyra/ip/access/epo/client.py
@@ -28,15 +28,29 @@ def includeme(config):
 
 def attach_oauth_client(event):
     #logger.info('Attaching OAuth client to request')
+
     request = event.request
     registry = request.registry
-    #context = request.context
 
     pool = registry.getUtility(IOpsOAuthClientPool)
+
+    # User-associated OPS credentials
     if request.user and request.user.upstream_credentials and request.user.upstream_credentials.has_key('ops'):
         request.ops_oauth_client = pool.get(request.user.userid, request.user.upstream_credentials['ops'])
+
+    # System-wide OPS credentials
     else:
-        request.ops_oauth_client = pool.get('default')
+        datasource_settings = registry.datasource_settings
+        datasources = datasource_settings.datasources
+        datasource = datasource_settings.datasource
+        if 'ops' in datasources and 'ops' in datasource and 'consumer_key' in datasource.ops and 'consumer_secret' in datasource.ops:
+            system_credentials = {
+                'consumer_key': datasource.ops.consumer_key,
+                'consumer_secret': datasource.ops.consumer_secret,
+            }
+            request.ops_oauth_client = pool.get('system', system_credentials)
+        else:
+            request.ops_oauth_client = pool.get('defunct')
 
 
 # ------------------------------------------
@@ -218,13 +232,13 @@ class OpsOAuthClientFactory(object):
     def __init__(self, credentials=None, debug=False):
 
         if credentials:
-            # User-associated OPS credentials
             self.consumer_key    = credentials['consumer_key']
             self.consumer_secret = credentials['consumer_secret']
+
         else:
-            # Elmyra OPS credentials
-            self.consumer_key    = r'***REMOVED***'
-            self.consumer_secret = r'***REMOVED***'
+            message = u'No credentials configured for OPS API.'
+            logger.error(u'OpsOAuthClientFactory: ' + message)
+            raise HTTPBadGateway(message)
 
         if debug:
             logging.getLogger('oauthlib').setLevel(logging.DEBUG)