+++ elmyra.ip.access.epo/CHANGES.rst
@@ -9,6 +9,7 @@ development
 - Improve exception handling for authentication errors against OPS API
 - Improve datasource configuration mechanics
 - Improve OPS API error propagation
+- Pull application-wide OPS API authentication credentials from datasource settings
 
 2017-09-08 0.135.0
 ==================